frysk.sys.Tid returns a ProcessIdentifier.
frysk-core/frysk/event/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* EventLoop.java: Update to match frysk.sys.Tid.
	* EventLoopTestBed.java: Ditto.

frysk-sys/frysk/sys/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* Tid.java: Return a ProcessIdentifier.

+++ frysk-core/frysk/proc/live/LinuxPtraceHost.java
@@ -257,7 +257,7 @@ public class LinuxPtraceHost extends LiveHost {
 		    ProcessIdentifier pid
 			= Fork.ptrace(stdin, stdout, stderr, args);
 		    // See if the Host knows about this task.
-		    TaskId myTaskId = new TaskId(Tid.get());
+		    TaskId myTaskId = new TaskId(Tid.get().intValue());
 		    Task myTask = get(myTaskId);
 		    if (myTask == null) {
 			// If not, find this process and add this task to it.
